Disable SSR to avoid hydration errors
